refactor(app_view): split Opt + dropdown option fns into opts.rs
`struct Opt`, its `impl Opt` / `impl SelectItem`, the
`make_select` / `read_select` wrappers, and every per-field
option-list builder (baud, data_bits, parity, stop_bits,
flow_control, line_ending, line_policy, backspace, port) move
into a new `app_view::opts` child module. About 200 lines, all
of which were a coherent profile-form helper layer.

`Opt` keeps its fields private; only the `pub(super) fn new`
constructor + the `SelectItem` impl reach across the module
boundary. Same descendants-see-parent rule used for the earlier
splits — opts.rs can call back into `super::AppView` for the
`Context<AppView>` it needs in `make_select` / `read_select`.

`port_opts` rides along because it returns the same `Vec<Opt>`
shape and the editor composes it alongside the static lists,
even though its data comes from `serial::ports::list_ports()`
rather than a hardcoded table. The `pub use ...::ports` in
mod.rs's import block drops out as a side effect — its only
remaining user is now in opts.rs.

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
